refactor: 상품 패키지 구조 변경
