+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -1,10 +1,11 @@
 package kitchenpos.application;
 
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuRepository;
-import kitchenpos.domain.Product;
-import kitchenpos.domain.ProductRepository;
-import kitchenpos.infra.PurgomalumClient;
+import kitchenpos.tobe.menu.domain.Menu;
+import kitchenpos.tobe.menu.application.port.out.MenuRepository;
+import kitchenpos.tobe.product.domain.Product;
+import kitchenpos.tobe.product.application.port.out.ProductRepository;
+import kitchenpos.tobe.external.port.out.PurgomalumClient;
+import kitchenpos.tobe.product.application.ProductService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
